Clean up implementation and fix return type

+++ src/Common/src/Interop/Unix/System.Private.CoreLib.Native/Interop.GetCpuUtilization.cs
@@ -19,6 +19,6 @@ internal static partial class Interop
         }
 
         [DllImport(Libraries.CoreLibNative, EntryPoint = "CoreLibNative_GetCpuUtilization")]
-        internal static extern unsafe IntPtr GetCpuUtilization(ref ProcessCpuInformation previousCpuInfo);
+        internal static extern unsafe int GetCpuUtilization(ref ProcessCpuInformation previousCpuInfo);
     }
 }
